add PassengerRequestScheduled/SubmittedEvent + handlers; keep DrtRequestSubmittedEvent as a subclass
